Move some tests arround.
