+++ packages/kartoffelgames.web.potato_web_builder/source/module/slot_instruction/slot-instruction-module.ts
