+++ packages/kartoffelgames.web.potato_web_builder/test/source/core/component/component.test.ts
@@ -1,19 +1,19 @@
 import { expect } from 'chai';
-import { Component, IComponentOnAttributeChange, IComponentOnDeconstruct, IComponentOnUpdate } from '../../source/core/component/component';
-import { ComponentRegister } from '../../source/core/component/component-register';
-import { PwbComponent } from '../../source/core/component/pwb-component.decorator';
-import { PwbConfiguration } from '../../source/core/configuration/pwb-configuration';
-import { CoreEntityProcessorProxy } from '../../source/core/core_entity/interaction-tracker/core-entity-processor-proxy';
-import { Processor } from '../../source/core/core_entity/processor';
-import { UpdateLoopError } from '../../source/core/core_entity/updater/update-loop-error';
-import { UpdateMode } from '../../source/core/enum/update-mode.enum';
-import { UpdateTrigger } from '../../source/core/enum/update-trigger.enum';
-import { IExpressionOnUpdate } from '../../source/core/module/expression_module/expression-module';
-import { PwbExpressionModule } from '../../source/core/module/expression_module/pwb-expression-module.decorator';
-import { PwbExport } from '../../source/module/export/pwb-export.decorator';
-import '../mock/request-animation-frame-mock-session';
-import '../utility/chai-helper';
-import { TestUtil } from '../utility/test-util';
+import { Component, IComponentOnAttributeChange, IComponentOnDeconstruct, IComponentOnUpdate } from '../../../../source/core/component/component';
+import { ComponentRegister } from '../../../../source/core/component/component-register';
+import { PwbComponent } from '../../../../source/core/component/pwb-component.decorator';
+import { PwbConfiguration } from '../../../../source/core/configuration/pwb-configuration';
+import { CoreEntityProcessorProxy } from '../../../../source/core/core_entity/interaction-tracker/core-entity-processor-proxy';
+import { Processor } from '../../../../source/core/core_entity/processor';
+import { UpdateLoopError } from '../../../../source/core/core_entity/updater/update-loop-error';
+import { UpdateMode } from '../../../../source/core/enum/update-mode.enum';
+import { UpdateTrigger } from '../../../../source/core/enum/update-trigger.enum';
+import { IExpressionOnUpdate } from '../../../../source/core/module/expression_module/expression-module';
+import { PwbExpressionModule } from '../../../../source/core/module/expression_module/pwb-expression-module.decorator';
+import { PwbExport } from '../../../../source/module/export/pwb-export.decorator';
+import '../../../utility/request-animation-frame-mock-session';
+import '../../../utility/chai-helper';
+import { TestUtil } from '../../../utility/test-util';
 
 describe('HtmlComponent', () => {
     before(() => {
