+++ packages/kartoffelgames.web.potato_web_builder/test/source/module/for-instruction.test.ts
@@ -3,11 +3,11 @@ import { PwbComponent } from '../../../source/core/component/pwb-component.decor
 import { PwbConfiguration } from '../../../source/core/configuration/pwb-configuration';
 import { Processor } from '../../../source/core/core_entity/processor';
 import { PwbExport } from '../../../source/module/export/pwb-export.decorator';
-import '../../mock/request-animation-frame-mock-session';
+import '../../utility/request-animation-frame-mock-session';
 import '../../utility/chai-helper';
 import { TestUtil } from '../../utility/test-util';
 
-describe('ForInstructionModule', () => {
+describe('ForInstruction', () => {
     before(() => {
         PwbConfiguration.configuration.updating.frameTime = Number.MAX_SAFE_INTEGER;
         PwbConfiguration.configuration.error.print = false;
